MethodResolution: enable boolean resolution
Adds the Java primitive class, Boolean.TYPE, to the conversion map in
MethodResolution. This allows equivalent boolean parameters in overloaded
methods to be appropriately considered assignable to each other.

Prior to this patch, overloaded methods "m1(boolean, Object...)" and
"m1(boolean, Map)" would be treated as unresolvable due to the perceived
inequality of boolean parameters.