jpype#1091: allow null characters in strings when convertStrings=True
Previously, `PyUnicode_FromString(str.c_str())` was used when
convertStrings=True. If the string contains null bytes, this
will terminate the string early. To allow strings with null
bytes to be returned, `PyUnicode_FromStringAndSize(str.c_str(), str.length())`
is now used instead, which create a Unicode object from the pointer
with the given length (instead of the position of the first
NULL character), thus allowing null bytes in the string.