[JPConversionAsBoolean] fixed numpy 2.3 breaking change of np.bool_ not being a Python bool anymore.
Historically, NumPy had its own boolean scalar type, numpy.bool_, which is a subclass of numpy.generic but not of Python’s built-in bool. For a long time it was also an instance of PyBool_Type — so PyBool_Check(obj) returned true for both True/False and for numpy.bool_ scalars.

Fixes #1307